+++ tools/generator-fluid/app/templates/jest-puppeteer.config.js
@@ -13,7 +13,7 @@ module.exports = {
   launch: {
     args: ['--no-sandbox', '--disable-setuid-sandbox'], // https://github.com/puppeteer/puppeteer/blob/master/docs/troubleshooting.md#setting-up-chrome-linux-sandbox
     dumpio: true, // output browser console to cmd line
-    slowMo: 100, // slows down process for easier viewing
-    headless: false, // run in the browser
+    // slowMo: 100, // slows down process for easier viewing
+    // headless: false, // run in the browser
   },
 };
