+++ tools/generator-fluid/package.json
@@ -26,6 +26,7 @@
     "@types/mocha": "^5.2.5",
     "make-promises-safe": "^5.1.0",
     "mocha": "^7.1.1",
+    "shelljs": "^0.8.4",
     "yeoman-assert": "^3.1.1",
     "yeoman-test": "^2.4.1"
   }
